added 'coming soon' to landing page to fix confusion - moved reusable components into folders for better organization
