--- src/components/Home.js
@@ -1,14 +1,0 @@
-import React from "react"
-import lambdaLandingUX from "../assets/LambdaLandingUX.png"
-
-const Home = () => {
-  return (
-    <img
-      className="landingMockup"
-      src={lambdaLandingUX}
-      alt="lambda landing page mockup"
-    />
-  )
-}
-
-export default Home
