+++ src/styles/navbar.scss
@@ -31,5 +31,12 @@
         
       }
     }
+    
+    label{
+      padding: 15px;
+      border: 1px solid $lambdaTeal;
+      border-radius: 20px 20px 20px 0 
+    }
+
   }
 }
